MAHOUT-294: Added arguments to the DistributedLanczosSolver CLI so that it can optionally launch the EigenVerificationJob. Added a new run() method with additional arguments needed by EVJ. Added new unit tests of the CLI and adjusted some of the cluster dumper tests to use the new run method. All tests run.
NOTE: this patch changes the semantics of the DLS --output argument: Formerly
it was the path to the output (raw) eigenvectors file. Now it is a path to an output
directory in which a rawEigenvectors file will be written. If the EVJ is also specified,
then the output directory will contain a cleanEigenvectors file in addition.

git-svn-id: https://svn.apache.org/repos/asf/mahout/trunk@997007 13f79535-47bb-0310-9956-ffa450edef68